backend/requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0

# Google Cloud - Essential only
google-cloud-bigquery==3.17.0
google-cloud-geminidataanalytics==0.4.0
google-generativeai==0.3.2
google-auth==2.23.4
google-api-python-client==2.108.0

# Anthropic & LangGraph
anthropic[vertex]>=0.40.0
langgraph>=0.0.50
langchain-core>=0.1.0

# Data & ML
altair==5.5.0
pandas==2.3.2
faker==22.0.0

# Web scraping
aiohttp==3.9.0
beautifulsoup4==4.12.0
lxml==5.1.0

# Optional: For V2 Research Agent advanced features (JS-heavy sites)
# playwright==1.40.0  # Uncomment if you need JavaScript rendering

# Environment
python-dotenv==1.0.0
